fix: move version text into sidebar to prevent overflow (#835)
