+++ src/ui/src/components/layout/sidebar.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useState } from "react";
 import YakOutline from "/logos/SQ-YakShaver-LogoIcon-Outline.svg?url";
 import logoImage from "/logos/YakShaver-Horizontal-Color-Darkmode.svg?url";
 import { IdentityServerAuthManager } from "../auth/IdentityServerAuthManager";
@@ -6,6 +7,22 @@ import { SettingsDialog } from "../settings/SettingsDialog";
 import { SidebarLink } from "../ui/sidebar-link";
 
 export default function Sidebar() {
+  const [appVersion, setAppVersion] = useState<string>("");
+  const [commitHash, setCommitHash] = useState<string>("");
+
+  useEffect(() => {
+    const fetchVersion = async () => {
+      try {
+        const info = await window.electronAPI.releaseChannel.getCurrentVersion();
+        setAppVersion(info.version);
+        setCommitHash(info.commitHash);
+      } catch (error) {
+        console.error("Failed to fetch app version information:", error);
+      }
+    };
+    fetchVersion();
+  }, []);
+
   return (
     <div className=" fixed top-0 left-0 w-[18rem] h-full bg-black/60 border-r border-white/25 flex flex-col gap-6 p-8 z-40">
       <h1>
@@ -26,6 +43,13 @@ export default function Sidebar() {
       <div className="relative bottom-0 mt-auto flex flex-col gap-3 left-0">
         <SettingsDialog />
         <IdentityServerAuthManager />
+        {(appVersion || commitHash) && (
+          <div className="text-[10px] text-white/30 font-mono break-all leading-tight">
+            {/* {appVersion && `v${appVersion} `}
+            {commitHash && `(${commitHash})`} */}
+            v2026.4.16-BUILD.186.ecfca55 (ecfa55adfb2dc871622ac3f03d78dd18bf0040)
+          </div>
+        )}
       </div>
     </div>
   );
